sage/requirements.txt
# Google AI & ADK
google-adk>=0.1.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0

# Google Workspace APIs
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# MCP
mcp>=1.0.0

# API layer
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Database (AlloyDB / PostgreSQL)
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
pgvector>=0.3.0
alembic>=1.13.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.3.0

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
